requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai==0.3.2
requests==2.31.0
google-ai-generativelanguage==0.4.0